backend/pyproject.toml
[project]
name = "ecg-pcg-backend"
version = "0.1.0"
description = "ECG/PCG心音心电数据管理平台后端"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "KexMeow", email = "kex@example.com" }]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "python-json-logger>=2.0.7",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "boto3>=1.34.0",
    "tencentcloud-sdk-python>=3.0.1068",
    # 信号处理 & 科学计算
    "numpy>=2.0.0",
    "scipy>=1.12.0",
    "neurokit2>=0.2.7",
    "matplotlib>=3.8.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "wfdb>=4.1.0",
    "soundfile>=0.12.0",
    "email-validator>=2.3.0",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "B", "C4", "D", "UP"]
ignore = ["E501", "D203", "D212"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
